refactor: moving the healthcare client
